Fixes
* Rename POCL_DEVICE_ADDRESS_BITS to HOST_DEVICE_ADDRESS_BITS
  the original variable was really only related to the host CPU
* Use POCL_DEVICE_ADDRESS_BITS instead of CL_DEVICE_ADDRESS_BITS
  as kernel library macro, since CL_DEVICE_ADDRESS_BITS conflicts
  with a value in CL/cl.h, making including pocl_types.h in C/C++
  code problematic
* CMake: update LLVM regex to match LLVM 7.1.0
* fix lib/kernel/printf* to use fixed-size types
* CMake: fix TCE flags to have address bits in one place
* update tests to these changes

+++ lib/kernel/printf_base.c
@@ -36,7 +36,7 @@
 /********************************************/
 
 #ifdef BUILD_WITH_LIBC
-int snprintf (char *str, unsigned long size, const char *format, ...);
+int snprintf (char *str, size_t size, const char *format, ...);
 #else
 
 #include "errol/errol.h"
